Apply strict(er) code-formatting (Prettier)...
... and some further minor TSLint related improvements. Swap out the
"tslint-config-standard" configuration and use "just" the recommended
ruleset extended with SonarTS (and Prettier).

Include ESLint (and related components; SonarJS and Prettier) and add a
configuration for the single JavaScript file in the "test/ws" folder :)
Somewhat of a preparation for fully switching over the ESLint in the
near future.

Remove the now unused "index.ts" file; superseeded by "main.ts".