Lookup bridge IP address in HueSocket constructor
It appears that otherwise _every, single_ UDP-datagram causes a DNS
lookup on the bridge's hostname; causing a sustained 180,000 DNS lookups
per hour while the fireplace is running... 🤯

Quickly hacked it in for now; needs some more work. Probably better to
do this on connect (which requires a slight rework/update of either
IHueSocketOptions or the connect method itself).