Deploy/distribute using a Docker container
Instead of a second/separate distributable NPM-package. Given my plans
with the fireplace (and how it's supposed to slot into the Home
Assistant setup) this seems a bit more flexible.

Works in basis, but for now the ".fireplace.json" file needs to be
mounted into the container for the script to run.

Should update to s6-overlay v3, for now hacked the v2 implementation
from tc66c-mqtt back to life (basically by adding the missing GPG-key
to the repository for now).